Introduce MetricsBucketToggle interface
