+++ src/Metrics/DefaultMetricsBucketToggle.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Unleash\Client\Metrics;
+
+use Unleash\Client\DTO\Feature;
+use Unleash\Client\DTO\Variant;
+
+/**
+ * @internal
+ */
+final class DefaultMetricsBucketToggle implements MetricsBucketToggle
+{
+    public function __construct(
+        private readonly Feature $feature,
+        private readonly bool $success,
+        private readonly ?Variant $variant = null,
+    ) {
+    }
+
+    public function getFeature(): Feature
+    {
+        return $this->feature;
+    }
+
+    public function isSuccess(): bool
+    {
+        return $this->success;
+    }
+
+    public function getVariant(): ?Variant
+    {
+        return $this->variant;
+    }
+}
