+++ src/Metrics/MetricsBucket.php
@@ -2,7 +2,18 @@
 
 namespace Unleash\Client\Metrics;
 
-interface MetricsBucket
-{
+use DateTimeInterface;
+use JsonSerializable;
 
+interface MetricsBucket extends JsonSerializable
+{
+    public function getStartDate(): DateTimeInterface;
+
+    public function setStartDate(DateTimeInterface $date): self;
+
+    public function getEndDate(): ?DateTimeInterface;
+
+    public function setEndDate(?DateTimeInterface $date): self;
+
+    public function addToggle(MetricsBucketToggle $toggle): self;
 }
