+++ src/Metrics/MetricsBucketToggle.php
@@ -5,30 +5,11 @@ namespace Unleash\Client\Metrics;
 use Unleash\Client\DTO\Feature;
 use Unleash\Client\DTO\Variant;
 
-/**
- * @internal
- */
-final class MetricsBucketToggle
+interface MetricsBucketToggle
 {
-    public function __construct(
-        private readonly Feature $feature,
-        private readonly bool $success,
-        private readonly ?Variant $variant = null,
-    ) {
-    }
+    public function getFeature(): Feature;
 
-    public function getFeature(): Feature
-    {
-        return $this->feature;
-    }
+    public function isSuccess(): bool;
 
-    public function isSuccess(): bool
-    {
-        return $this->success;
-    }
-
-    public function getVariant(): ?Variant
-    {
-        return $this->variant;
-    }
+    public function getVariant(): ?Variant;
 }
